requirements.txt
requests
jsonschema
pyyaml
